requirements.txt
streamlit~=1.50.0
python-dotenv>=1.0.0
google-genai>=0.3.0
requests~=2.32.5

GitPython~=3.1.45
gitdb~=4.0.12
numpy~=2.3.4
pillow~=11.3.0
pip~=25.0.1
typing_extensions~=4.15.0
attrs~=25.4.0
protobuf~=6.33.0
pytz~=2025.2
toml~=0.10.2
smmap~=5.0.2
click~=8.3.0
idna~=3.11
charset-normalizer~=3.4.4
altair~=5.5.0
jsonschema~=4.25.1
narwhals~=2.9.0
pandas~=2.3.3
pyarrow~=21.0.0
Jinja2~=3.1.6
packaging~=25.0
referencing~=0.37.0
cachetools~=6.2.1
certifi~=2025.10.5
urllib3~=2.5.0
tenacity~=9.1.2
MarkupSafe~=3.0.3
python-dateutil~=2.9.0.post0
pydeck~=0.9.1
tzdata~=2025.2
tornado~=6.5.2
colorama~=0.4.6
six~=1.17.0
watchdog~=6.0.0
blinker~=1.9.0
jsonschema-specifications~=2025.9.1
google-genai
google-generativeai